backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pillow==10.3.0
torch==2.4.0
torchvision==0.19
numpy==1.26.4
